refactor(kernel): authorize staged domain dispatch
Move native start context and dispatch primitives into a dedicated typed manager module. Authorize verifier-owned staged state before the sole Running transition and add focused dispatch authorization regressions.

Tracking: #1766
