+++ changes/1766.changed.md
@@ -1,0 +1,6 @@
+Native domain start context is staged while a domain remains Prepared and
+carried privately with its verified manifest, component, and handle tuple.
+Dispatch revalidates that authorization and state before its sole transition
+to Running. This is staged-start return evidence only, not readiness, a
+Running stop, publication, or a production returning-lifecycle oracle.
+Refs #1766
