+++ changes/1766.fixed.md
@@ -1,9 +1,6 @@
 Ring 0 consumes one verifier-owned admitted generation after canonical
 manifest verification. Component admission and each native domain handle
-generation derive from that same manifest/component tuple. Manager context
-validation now returns a private staged object while the domain remains
-Prepared; dispatch alone revalidates that triad and state before its sole
-transition to Running. Reclamation marks the adapter record stale, and
-substituted admission identities fail closed. This is staged-start return
-evidence only, not readiness, a Running stop, publication, or a production
-returning-lifecycle oracle. Refs #1766
+generation derive from that same manifest/component tuple; reclamation marks
+the adapter record stale, and substituted admission identities fail closed.
+This is identity-retention and admission evidence only, not returning
+lifecycle or publication evidence. Refs #1766
